+++ src/commands/createModule.ts
@@ -1,12 +1,10 @@
 import {
-  API_TEMPLATE,
-  CONSTANTS_TEMPLATE,
-  MAIN_IMPORT_TEMPLATE,
-} from "../constants/index.js";
-import {
-  capitalizeComponentName,
+  capitalizeComponentPath,
+  capitalizeFirstLetter,
   createDirectoryIfNotExists,
   getComponentsPaths,
+  getTemplateContentWithName,
+  uncapitalizeFirstLetter,
   writeFile,
 } from "../helpers/index.js";
 import { createComponent } from "./createComponent.js";
@@ -15,19 +13,21 @@ export function createModule(
   moduleName: string,
   options: { full?: boolean; startComponent?: string }
 ) {
-  const modulePath = moduleName
-    .split("/")
-    .map(capitalizeComponentName)
-    .join("/");
-  const module = capitalizeComponentName(moduleName.split("/").pop());
+  const modulePath = capitalizeComponentPath(moduleName);
+  const capitalizedModuleName = capitalizeFirstLetter(
+    moduleName.split("/").pop()
+  );
+  const uncapitalizedModuleName = uncapitalizeFirstLetter(
+    moduleName.split("/").pop()
+  );
   const startComponent = options.startComponent
-    ? capitalizeComponentName(options.startComponent)
-    : module;
+    ? capitalizeComponentPath(options.startComponent)
+    : capitalizedModuleName;
 
   const paths = getComponentsPaths(`src/modules/${modulePath}`, {
     mainModule: "index.ts",
     api: "api",
-    apiFile: `api/${module.charAt(0).toLowerCase() + module.slice(1)}Api.ts`,
+    apiFile: `api/${uncapitalizedModuleName}Api.ts`,
     constants: "constants",
     constantFile: "constants/index.ts",
     hooks: "hooks",
@@ -39,9 +39,18 @@ export function createModule(
   createDirectoryIfNotExists(paths.baseDir);
 
   // Templates for the files
-  const apiTemplate = API_TEMPLATE(module);
-  const constantsTemplate = CONSTANTS_TEMPLATE;
-  const mainImportTemplate = MAIN_IMPORT_TEMPLATE(startComponent);
+  const apiTemplate = getTemplateContentWithName(
+    "api.ts",
+    uncapitalizedModuleName
+  );
+  const constantsTemplate = getTemplateContentWithName(
+    "constants.ts",
+    capitalizedModuleName
+  );
+  const mainImportTemplate = getTemplateContentWithName(
+    "mainImport.ts",
+    startComponent
+  );
 
   if (options?.full) {
     createDirectoryIfNotExists(paths.api);